Resolve "Link favourites and watchlist items to users"

+++ app/src/contexts/moviesContext.jsx
@@ -1,24 +1,48 @@
-import React, { useState } from "react";
+import React, { useState, useEffect } from "react";
 import { getUserSelection } from "../api/supabase";
+import { useAuth } from "../hooks/useAuth";
 import { supabaseClient } from "../utils/client";
 
 export const MoviesContext = React.createContext(null);
 
-const userFavourites = await getUserSelection("favourites");
-const userWatchlist = await getUserSelection("watchlist");
-
 const MoviesContextProvider = (props) => {
-  const [favourites, setFavourites] = useState(userFavourites);
-  const [watchlist, setWatchlist] = useState(userWatchlist);
-  const [myReviews, setMyReviews] = useState( {} )
+  const { session }  = useAuth();
+  const [favourites, setFavourites] = useState( [] );
+  const [watchlist, setWatchlist] = useState( [] );
+  const [myReviews, setMyReviews] = useState( {} );
 
-  const addToFavourites = async (movie) => {
+  useEffect(() => {
+    //Setup favourites and watchlist from user data present in database
+    getUserFavourites(session?.user?.id);
+    getUserWatchlist(session?.user?.id);
+  }, [session]);
+
+  // Get users favourites from database, populate server state
+  const getUserFavourites = async (user) => {
+    let userFavourites = await getUserSelection("favourites", user);
+    if(userFavourites === undefined){
+      setFavourites([]);
+    }else{
+      setFavourites(userFavourites);
+    }
+  }
+
+  // Get users watchlist items from database, populate server state
+  const getUserWatchlist = async (user) => {
+    let userWatchlist = await getUserSelection("watchlist", user);
+    if(userWatchlist === undefined){
+      setWatchlist([]);
+    }else{
+      setWatchlist(userWatchlist);
+    }
+  }
+
+  const addToFavourites = async (movie, user) => {
     try{
         let updatedFavourites = [...favourites];
         if(!favourites.includes(movie.id)) {
           // add to supabase DB table "favourites"
-          let { error } = await supabaseClient
-            .from("favourites").insert({id: movie.id});
+          let { error } = await supabaseClient.from("favourites").insert({id: movie.id, user_id: user});
           // if no error returned from supabase, add to server state
           if(!error){
             updatedFavourites.push(movie.id);
@@ -32,30 +56,25 @@ const MoviesContextProvider = (props) => {
     }
   }
 
-  const removeFromFavourites = async (movie) => {
+  const removeFromFavourites = async (movie, user) => {
     try{
-      let { error } = await supabaseClient
-        .from("favourites").delete().eq("id", movie.id);
-
+      let { error } = await supabaseClient.from("favourites").delete().eq("id", movie.id).eq("user_id", user);
       if(!error){
         //remove from server state
         setFavourites(favourites.filter((mId) => mId !== movie.id));
       }else{
         throw error;
       }
-    
     }catch(err){
       console.log(err);
     }
   };
 
-  const addToWatchlist = async (movie) => {
+  const addToWatchlist = async (movie, user) => {
     try{
       let updatedWatchlist = [...watchlist];
       if(!watchlist.includes(movie.id)) {
-        let { error } = await supabaseClient
-        .from("watchlist").insert({id: movie.id});
-      
+        let { error } = await supabaseClient.from("watchlist").insert({id: movie.id, user_id: user});
         if(!error){
           updatedWatchlist.push(movie.id)
         }else{
@@ -68,10 +87,9 @@ const MoviesContextProvider = (props) => {
     }
   }
 
-  const removeFromWatchlist = async (movie) => {
+  const removeFromWatchlist = async (movie, user) => {
     try{
-      let { error } = await supabaseClient
-        .from("watchlist").delete().eq("id", movie.id);
+      let { error } = await supabaseClient.from("watchlist").delete().eq("id", movie.id).eq("user_id", user);
       
       if(!error){
         setWatchlist(watchlist.filter((mId) => mId !== movie.id));
@@ -93,6 +111,8 @@ const MoviesContextProvider = (props) => {
       value={{
         favourites,
         watchlist,
+        getUserFavourites,
+        getUserWatchlist,
         addToFavourites,
         removeFromFavourites,
         addToWatchlist,
